Add tests for analyzing web page
Add tests for the whole DynamicPageAnalyzer pipeline: opening
a web page given it's URL, running Analyzer on it.
For that, a small module for running web server serving test files
by HTTP is added, it works the same way as the one used by
SlimerJS own tests.
Currently, a few basic tests are added.