Test runner: choose which tests to run
This commit adds an ability for test runner to configure which
tests to run. The main reason for this change is that set of tests
runnable on NodeJS and on SlimerJS are now not equal: SlimerJS can
run tests for analyzing web page, but NodeJS can't. So, now test
runner checks whether it's running on SlimerJS and adds page tests
only in that case, otherwise they are excluded. To protect from
silently skipping slimer-specific tests when they are desired, flag
'--ensure-slimer' is added that allows to ensure that tests are
running under SlimerJS. If it is given and runner thinks it is not
running on Slimer, it will fail with error.
Additionally, an ability to include only particular test file/set
of test files is added: all command line arguments for test runner
except '--ensure-slimer' are considered as test file patterns
(relative to 'test/analyzer' dir), which will override default set
of test file patterns.
Info about this test runner usage is added to README.