Headless bot: wait more for a page to load
If all requests that were made from a page are completed, it does
not mean the page is fully loaded. It may be that after the last
request is completed, JS code on a page will make some more
requests to server, loading more resources - effectively
performing next stage of page loading. To handle that, wait a bit
(400 msec) and, if no requests are made during that period,
consider a page to be loaded. If there were more requests during
that period, start over.