Record stacks for other requests besides XHR/fetch
If a request is not sent via XHR or fetch, but is triggered by JS
(for example, by setting .src on Image or by adding a script with
src to DOM), we may still be able to get it's stacktrace because
it usually grows from the point in client code that triggered the
request. So we now record stack snapshot in request handler for
all requests (except main "document" request), not only for XHR
and fetch.