Fetch model: replace type cast with type checks
Fetch settings were explicitly type-casted to type FetchSettings.
(This was forcibly setting field types for TypeScript, even when
real types could differ)
Remove that cast and check types (sometimes converting values)
instead. This fixes a bug when method (that should be a string)
turned out to be of type Unknown, leading to and error caused by
calling a non-existent method.