ValueSet: allow nonfatal invalid string conversion
ValueSet should not be cast to a primitive type except a few
special cases (debugging prints in analyzer itself for example).
The best way to spot invalid unexpected conversions to string is
to crash with an exception, but, to be more robust, add code to
issue a warning instead, and to convert to 'UNKNOWN'.