base tag tests: rename 1 test and add another one
Add a test with 'relative base url' so that URL in base tag
in it is really relative (does not start with a '/' or a scheme).
Also, rename old 'relative base url' test to emphasize that
it has not slash at the end of the URL (not at the beginning).