Support getting event handler attrs code on Chrome
Add support for method triggerParsingOfEventHandlerAttributes()
under Chrome. For Chrome it's a little bit more tricky than under
Firefox, because arrival of handler source code to debugger is
not synchronous and we have to wait for it to happen somehow,
otherwise we will miss some handler attrs because their
'Debugger.scriptParsed' event arrived later than we started code
analysis.