Explicitly depend on 'punycode' module
This module is builtin in NodeJS, but in SlimerJS we need it as
there is no such module neither in SlimerJS itself, nor in our
shims. It is sometimes installed accidentally as a transitive
dependency, but sometimes not. Besides, the builtin version is
marked as deprecated in NodeJS and they recommend to use the NPM
package.