Match functions by values when building call chain
Functions were previously identified using variable bindings.
So, only free-standing functions assigned to variables could
participate in call chains. They are now identified by function
values - so, if analyzer is able to determine function value at
call site, it should be able to build a call chain through it.

This commit also removes an 'optimization': if we were building
a call chain with a call to function X inside function Y and Y
contained multiple calls to X, these cases were 'collapsed' and
we built only one chain. There was no research about the impact of
this optimization, so, now, as code changed and it became harder
to maintain it, it is removed. If we discover we really need it,
we will re-introduce it later.