Change assignment of prop with ValueSet name
Until now, if name of a prop being assigned was a ValueSet, such
an assignment was cancelled. Now, analyzer tries to pick one
element from the set and use it as a name for the prop. The name
being picked is chosen by prioritizing values and picking one
with highest priority. There are 3 kinds of priority:
  - 'Unknown' values have the lowest priority
  - 'null' and 'undefined' have the average priority
  - all remaining values have the highes priority
Choice between values with equal priority is made arbitrarily