Push empty args for module function
Args are pushed to integrate better with other mechanisms in
analyzer that expect that if a function is in AST, arg set will be
pushed to .argsStack. Among such mechanisms is popping args from
stack for every function. Arg set is empty because we don't want
arg-searching mechanisms to search for args of module function.