Remove 'unknown' from Value type definition
Type definition for the Value type (in types/generic.ts) contained
an variant with 'unknown' type, because otherwise TypeScript
failed with an error complaining that a type references itself.
Current TypeScript version is actually capable of handling
recursive types, in our case it is possible if we use object
literal notation instead of 'Record' template for the 'object'
variant. Apply this fix and get rid of 'unknown' type in Value
type definition.