Enter function calls
When determining function return value, analyzer now attempts to
enter the body of that function, using actual arguments and
analyzing what happens inside. Call entering depth is limited,
currently depth limit is set to 2.
For now, calls with multiple possible callees (when called
function is a ValueSet) are not supported